+++ test/integration/run.sh
@@ -65,14 +65,17 @@ TESTS=(
 	"test_notification_dedup.sh"
 	"test_cancellation.sh"
 	"test_cli_guards.sh"
-	"test_cli_init_scaffold_server.sh"
+	"test_cli_init_new.sh"
 	"test_cli_init_config_doctor.sh"
 	"test_cli_config_variants.sh"
 	"test_cli_validate_fix.sh"
 	"test_cli_validate_errors.sh"
 	"test_project_root_detection.sh"
 	"test_elicitation.sh"
+	"test_elicitation_modes.sh"
+	"test_icons.sh"
 	"test_cli_scaffold_test.sh"
+	"test_scaffold_smoke.sh"
 )
 
 passed=0